requirements.txt
requests
gradio-client
python-telegram-bot
